requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-text-splitters
langchain-chroma
langchain-ollama
langchain-groq
langchain-community
pypdf